pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "~=3.12"
dependencies = [
    "fastapi>=0.111.0,<0.112",
    "uvicorn[standard]>=0.29.0,<0.30",
    "pydantic[email]>=2.7.1,<3",
    "pydantic-settings>=2.2.1,<3",
    "sqlalchemy[asyncio]>=2.0.30,<3",
    "asyncpg>=0.29.0,<0.30",
    "alembic>=1.13.1,<2",
    "gunicorn>=23.0.0,<24",
    "passlib>=1.7.4,<2",
    "orjson>=3.10.12,<4",
    "redis>=5.2.0,<6",
    "python-jose[cryptography]>=3.3.0,<4",
    "fastapi-users-db-sqlalchemy>=6.0.1,<7",
    "pillow>=11.1.0,<12",
    "prometheus-client>=0.21.1",
]

[dependency-groups]
dev = [
    "black>=24.4.2,<25",
    "isort>=5.13.0,<6",
    "ruff>=0.4.0,<1",
]

[tool.uv]
package = false

# Black - форматтер кода
[tool.black]
line-length = 88
target-version = ['py312']

# isort - сортировка импортов
[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app_src", "api", "core", "crud", "utils"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# Ruff - быстрый линтер
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

